update as per PR review
